combine samples to test
